feat: healtcheck and version support
The oslc-request-server binary now supports a `--version` flag.

Additionally a new `healthcheck` command has been added. This command will use the grpc configuration options, same as is used to setup the server, to determine which server to connect to.

In order to detect if the healthcheck was successful, the exit code should be referenced. It does output to stdout as well, but the exact output should not be relied on.